Remove SingleValueDocValuesFieldUpdates abstract class
